Update Svelte wrapper links in README.md
I removed [gtm-nayan/svelte-use-chessground](https://github.com/gtm-nayan/svelte-use-chessground) as it has been archived with not commits in the last 4 years.
Also for clarity to the reader, I added labels for which Svelte version each wrapper is built for. This may also clear confusion on why they are three of them.